WIP: Circular import errors, looking at cause
